Commit du jour

+++ Teensy/lib/LesInseparables/Codeuse.h
@@ -42,7 +42,7 @@ class Codeuse
     int32_t oldTicks; //Nombre de ticks compté, nombre de ticks compté au dernier appel
     float debug;             //Distance parcourue par la codeuse depuis l'alllumage
     float diametreRoue;
-    uint16_t ticksPerRound; //Nombre de ticks par tours de roue
+    double ticksPerRound; //Nombre de ticks par tours de roue
     Encoder *enc;           //Objet Encoder de la librairie Encoder.
 
 public:
@@ -50,7 +50,7 @@ public:
     int32_t ticks;
     void actuate(float dt); //Actualise (transforme les ticks en vitesse puis en avance)
     Codeuse();
-    Codeuse(uint8_t pinA, uint8_t pinB, uint16_t ticksPerRound, float diametreRoue);
+    Codeuse(uint8_t pinA, uint8_t pinB, double ticksPerRound, float diametreRoue);
 };
 
 /**
@@ -68,7 +68,7 @@ public:
     Codeuse codeuseGauche, codeuseDroite;
     void updateCinetique(float dt);
     Odometrie();
-    Odometrie(uint16_t ticksPerRound, Cinetique *cinetique, float eloignementCodeuses,
+    Odometrie(double ticksPerRound, Cinetique *cinetique, float eloignementCodeuses,
               uint8_t pinACodeuseGauche, uint8_t pinBCodeuseGauche, float diametreRoueGauche,
               uint8_t pinACodeuseDroite, uint8_t pinBCodeuseDroite, float diametreRoueDroite,
               uint8_t pinInterDroite, uint8_t pinInterGauche);
